README.md
Summarize the project and what problem it was solving.
The goal of this project was to create a user dashboard to aid in locating and sorting pets available for adoption in the Austin, TX area.
What did you do particularly well?
I feel that the thing I excelled at in this project was re-factoring and finishing the product at the end.  I had difficulties earlier in implementation but was able to make everything work cleanly by revisiting the foundations of the program.
Where could you improve?
I think the display of the dashboard could be improved or maybe include functionality to support additional user logins.
What tools and/or resources are you adding to your support network?
Through the development of this project, I have added MongoDB as well as full-stack dev experience.
What skills from this project will be particularly transferable to other projects and/or coursework?
Any projects related to database management, python development or full-stack development use skills learned from this project.
How did you make this project maintainable, readable, and adaptable?
The use of best practices and well-commented code add to the maintainability, readability, and adaptability of this project.